requirements.txt
pdfplumber
python-docx
pytesseract
pillow
spacy
keybert
sentence-transformers
transformers
streamlit